[IMP] partner_external_map: Better layout + icon + some refactoring
