+++ partner_external_map/README.rst
@@ -6,47 +6,52 @@
 Partner External Maps
 =====================
 
-In the old days of Odoo/OpenERP, back in version 6.1, there was an official *google_map* module ; this module added a *Map* button on the partner form view and, when the user clicked on that button, it would open a new tab on its web browser and go to Google Map with a search on the address of the partner.
+In the old days of Odoo/OpenERP, back in version 6.1, there was an official
+*google_map* module ; this module added a *Map* button on the partner form view
+and, when the user clicked on that button, it would open a new tab on its web
+browser and go to Google Map with a search on the address of the partner.
 
 This module aims at restoring this feature with several improvements:
 
-* each user can select the map website he wants to use in its preferences
-
-* there are now two buttons on the partner form view: one to open a regular map on the address of the partner, and another one to open an itinerary map from the start address configured in the preferences of the user to the address of the partner.
+* Each user can select the map website he wants to use in its preferences
+* There are now two buttons on the partner form view: one to open a regular map
+  on the address of the partner, and another one to open an itinerary map from
+  the start address configured in the preferences of the user to the address of
+  the partner.
 
 This module supports several map websites:
 
 * `Google Maps <https://www.google.com/maps>`
-
 * `OpenStreetMap <https://www.openstreetmap.org/>`
-
 * `Bing Maps <https://www.bing.com/maps/>`
-
 * `Here Maps <https://www.here.com/>`
-
 * `MapQuest <http://www.mapquest.com/>`
-
 * `Yahoo! Maps <https://maps.yahoo.com/>`
 
-If the module *base_geolocalize* from the official addons is installed on the system, it will use the latitude and longitude to localize the partner (instead of the address) if this information is present on the partner.
+If the module *base_geolocalize* from the official addons is installed on the
+system, it will use the latitude and longitude to localize the partner (instead
+of the address) if this information is present on the partner.
 
 Configuration
 =============
 
-If you want to create additionnal map websites, go to the menu *Sales > Configuration > Address Book > Localization > Map Websites*. You are invited to send the configuration information of your additionnal map websites to the author of the module, so that the module can be updated with more pre-configured map websites.
+If you want to create additionnal map websites, go to the menu
+*Sales > Configuration > Address Book > Localization > Map Websites*. You are
+invited to send the configuration information of your additionnal map websites
+to the author of the module, so that the module can be updated with more
+pre-configured map websites.
 
 Usage
 =====
 
 First, you need to configure in your preferences:
 
-* the map website to use for the regular maps,
+* The map website to use for the regular maps,
+* The map website to use for the route maps,
+* The start address for the route maps.
 
-* the map website to use for the route maps,
-
-* the start address for the route maps.
-
-Then you can use the two new buttons on the partner form to open a regular map or a route map.
+Then you can use the two new buttons on the partner form to open a regular map
+or a route map.
 
 .. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
    :alt: Try me on Runbot
@@ -61,7 +66,7 @@ check there if your issue has already been reported. If you spotted it first,
 help us smashing it by providing a detailed and welcomed `feedback
 <https://github.com/OCA/
 partner-contact/issues/new?body=module:%20
-partner_external_maps%0Aversion:%20
+partner_external_map%0Aversion:%20
 9.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Credits
@@ -71,6 +76,7 @@ Contributors
 ------------
 
 * Alexis de Lattre <alexis.delattre@akretion.com>
+* Pedro M. Baeza <pedro.baeza@tecnativa.com>
 
 Maintainer
 ----------
