+++ partner_external_map/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
-from . import partner_external_maps
-from .post_install import set_default_map_settings
+from . import models
+from .hooks import set_default_map_settings
