+++ partner_external_map/models/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+from . import map_website
+from . import res_partner
+from . import res_users
