+++ partner_external_map/views/res_partner_view.xml
