+++ partner_external_map/views/res_users_view.xml
